Alterando tipo do valor de entrada para decimal e corrigendo erros relacionados

+++ Desafio2/ConversorMonetario/EntradaDeDados.cs
@@ -29,10 +29,10 @@
         /// Realiza a leitura do <see cref="ValorDeEntrada"/>.
         /// </summary>
         /// <returns>Retorna um <see langword="double"/> inserido no console.</returns>
-        public static double? ValorDeEntrada()
+        public static decimal? ValorDeEntrada()
         {
             Console.WriteLine("Valor: ");
-            return double.Parse(Console.ReadLine());
+            return decimal.Parse(Console.ReadLine());
         }
     }
 }
